refactor: share chatgpt codex usage limit formatter
