+++ turbo/apps/web/src/lib/zero/chat-thread/chat-run-error-message.ts
@@ -1,4 +1,7 @@
-import { RUN_ERROR_GUIDANCE } from "@vm0/api-contracts/contracts/errors";
+import {
+  formatChatgptCodexUsageLimitError,
+  RUN_ERROR_GUIDANCE,
+} from "@vm0/api-contracts/contracts/errors";
 import { asc, eq } from "drizzle-orm";
 import { agentRuns } from "@vm0/db/schema/agent-run";
 import { zeroRuns } from "@vm0/db/schema/zero-run";
@@ -8,8 +11,6 @@ const REPORT_ERROR_STREAK_THRESHOLD = 2;
 const CHAT_RUN_TRANSIENT_ERROR_MESSAGE =
   "Oops, something went wrong. Please try again later.";
 const CHAT_RUN_REPORTABLE_ERROR_MESSAGE = "An unexpected error occurred.";
-const CHATGPT_CODEX_USAGE_DETAILS_URL =
-  "https://chatgpt.com/codex/settings/usage";
 
 const ACTIONABLE_ERROR_SNIPPETS = [
   ...Object.values(RUN_ERROR_GUIDANCE).flatMap((guidance) => {
@@ -29,48 +30,6 @@ function isActionableRunError(errorMessage: string): boolean {
   return ACTIONABLE_ERROR_SNIPPETS.some((snippet) => {
     return normalized.includes(snippet.toLowerCase());
   });
-}
-
-function isChatgptCodexUsageLimitError(errorMessage: string): boolean {
-  const normalized = errorMessage.toLowerCase();
-  return (
-    normalized.includes("usage limit") &&
-    (normalized.includes("chatgpt.com/codex") ||
-      normalized.includes("codex/settings/usage") ||
-      normalized.includes("chatgpt codex"))
-  );
-}
-
-function extractChatgptCodexRetryPhrase(errorMessage: string): string | null {
-  const match = /\btry again\s+(at|after|in)\s+([^.;\n\r]+)/i.exec(
-    errorMessage,
-  );
-  if (!match) {
-    return null;
-  }
-
-  const preposition = match[1];
-  const retryAt = match[2]?.trim();
-  if (!preposition || !retryAt) {
-    return null;
-  }
-  if (retryAt.length > 80 || !/^[a-z0-9\s:,+/-]+$/i.test(retryAt)) {
-    return null;
-  }
-
-  return `${preposition.toLowerCase()} ${retryAt}`;
-}
-
-function formatChatgptCodexUsageLimitError(
-  errorMessage: string,
-): string | null {
-  if (!isChatgptCodexUsageLimitError(errorMessage)) {
-    return null;
-  }
-
-  const retryPhrase = extractChatgptCodexRetryPhrase(errorMessage);
-  const retrySentence = retryPhrase ? ` This limit resets ${retryPhrase}.` : "";
-  return `ChatGPT Codex usage limit reached.${retrySentence} View details in [ChatGPT Codex usage settings](${CHATGPT_CODEX_USAGE_DETAILS_URL}), or switch to another model to continue now.`;
 }
 
 function buildReportableErrorMessage(runId: string): string {
