+++ turbo/packages/api-contracts/src/contracts/errors.ts
@@ -115,3 +115,48 @@ export const RUN_ERROR_GUIDANCE: Record<
       "Wait for your current run to complete before starting a new one.",
   },
 };
+
+const CHATGPT_CODEX_USAGE_DETAILS_URL =
+  "https://chatgpt.com/codex/settings/usage";
+
+function isChatgptCodexUsageLimitError(errorMessage: string): boolean {
+  const normalized = errorMessage.toLowerCase();
+  return (
+    normalized.includes("usage limit") &&
+    (normalized.includes("chatgpt.com/codex") ||
+      normalized.includes("codex/settings/usage") ||
+      normalized.includes("chatgpt codex"))
+  );
+}
+
+function extractChatgptCodexRetryPhrase(errorMessage: string): string | null {
+  const match = /\btry again\s+(at|after|in)\s+([^.;\n\r]+)/i.exec(
+    errorMessage,
+  );
+  if (!match) {
+    return null;
+  }
+
+  const preposition = match[1];
+  const retryAt = match[2]?.trim();
+  if (!preposition || !retryAt) {
+    return null;
+  }
+  if (retryAt.length > 80 || !/^[a-z0-9\s:,+/-]+$/i.test(retryAt)) {
+    return null;
+  }
+
+  return `${preposition.toLowerCase()} ${retryAt}`;
+}
+
+export function formatChatgptCodexUsageLimitError(
+  errorMessage: string,
+): string | null {
+  if (!isChatgptCodexUsageLimitError(errorMessage)) {
+    return null;
+  }
+
+  const retryPhrase = extractChatgptCodexRetryPhrase(errorMessage);
+  const retrySentence = retryPhrase ? ` This limit resets ${retryPhrase}.` : "";
+  return `ChatGPT Codex usage limit reached.${retrySentence} View details in [ChatGPT Codex usage settings](${CHATGPT_CODEX_USAGE_DETAILS_URL}), or switch to another model to continue now.`;
+}
